Version 1.3.1-beta 1.0.3
Changes:
- Work has begun on rewriting the work with versions. The version will have to be stored in the app.version file, and not taken from the executable file, which should significantly improve the process of updating applications, especially whose metadata does not contain version information ("FileVersion" or "ProductVersion")
- Improved messages and their purpose
- Fixed bugs during deletion of temporary and unnecessary files

+++ src/main_func.rs
@@ -1,7 +1,17 @@
 use execute::Execute;
 use std::fs;
+use std::fs::File;
+use std::io::Write;
+use std::path::Path;
 use std::process;
 use std::process::Command;
+
+// Writing version information to file
+pub fn set_new_version(version: &str) {
+    let path = "app.version";
+    let mut file = File::create(path).expect("Error creating file with version information!");
+    file.write_all(version.as_bytes()).expect("Error writing version information to file!");
+}
 
 // Checking the Internet connection
 pub fn test_iconnection() -> Result<(), isahc::Error> {
@@ -42,19 +52,19 @@ pub fn task_kill(application_exe: &str) {
     let mut command = Command::new(TASKKILL_TOOL);
 
     command.arg("/F").arg("/T").arg("/IM").arg(application_exe);
-
-    let output = command.execute_output().unwrap();
-    if let Some(exit_code) = output.status.code() {
-        if exit_code == 0 {
-            println!("The application processes are killed. (this is good)");
-        } else {
-            eprintln!("The process was not found. (this is good)");
-        }
-    } else {
-        eprintln!("Interrupted!");
-        press_btn_continue::wait("Press any key to exit...").unwrap();
-        process::exit(1);
-    }
+    command.execute();
+    // let output = command.execute_output().unwrap();
+    // if let Some(exit_code) = output.status.code() {
+    //     if exit_code == 0 {
+    //         println!("The application processes are killed. (this is good)");
+    //     } else {
+    //         eprintln!("The process was not found. (this is good)");
+    //     }
+    // } else {
+    //     eprintln!("Interrupted!");
+    //     press_btn_continue::wait("Press any key to exit...").unwrap();
+    //     process::exit(1);
+    // }
 }
 
 // Downloading github release by redl
@@ -139,11 +149,19 @@ pub fn extracting(current_dir: &str) {
 // Delete portable installer
 pub fn delete_file(current_dir: &str, is_leave_folders: &bool) {
     let file_dir = String::from(format!("{}app.dat", current_dir));
-    fs::remove_file(file_dir).expect("Temporary file \"app.dat\" not found.");
+    if Path::new(&file_dir).exists() {
+        fs::remove_file(file_dir).expect("Temporary file \"app.dat\" not found.");
+    }
     if !is_leave_folders {
-        fs::remove_dir_all(format!("{}..\\$PLUGINSDIR", current_dir))
-            .expect("Unnecessary folder \"$PLUGINSDIR\" was not found.");
-        fs::remove_dir_all(format!("{}..\\Other", current_dir))
-            .expect("Unnecessary folder \"Other\" was not found.");
+        let mut dir  = format!("{}..\\$PLUGINSDIR", current_dir).to_string();
+        if Path::new(&dir).exists() {
+            fs::remove_dir_all(dir)
+                .expect("Unnecessary folder \"$PLUGINSDIR\" was not found.");
+        }
+        dir  = format!("{}..\\Other", current_dir).to_string();
+        if Path::new(&dir).exists() {
+            fs::remove_dir_all(dir)
+                .expect("Unnecessary folder \"Other\" was not found.");
+        }
     }
 }
